pyproject.toml
[tool.poetry]
name = "alpha"
version = "0.1.0"
description = "This is my first ever Generative AI project."
authors = ["Wahab Awudu <wahabawudu2020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.3.0"
requests = "^2.31.0"
python-decouple = "^3.8"
astrapy = "^0.6.0"
cassio = "^0.1.3"
datasets = "^2.15.0"
langchain = "^0.0.336"
tiktoken = "^0.5.1"
black = "^23.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
